Bug fix for some circumstances when an error is thrown when the dateProvider of a target for an AccordionLayout is changed.
AccordionLayout - If the seletedIndex is bigger than the length of the number of elements included in the layout, measures the item at index 0.

NavigatorLayoutBase - checks whether the selectedIndex is bigger than the number of items in the layout. If so, sets the proposed selectedIndex to 0. Also makes use of the indicesInLayout array to get the selectedElement.

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/whiteboard@1389146 13f79535-47bb-0310-9956-ffa450edef68