Implemented an alternate method to deal with default values for parameters in JS.
In order to do this without 'breaking' any existing code, I've create a new 'setting' on JSSharedData: OUTPUT_ALTERNATE. I used this setting to split the flow in 'emitFunctionBlockHeader' and call to 'emitDefaultParameterCodeBlock_Alternate' instead of the regular method.

I've also added a test to 'TestGoogEmiter': 'testDefaultParameter_Alternate' to check the functionality of the above mentioned addition.

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/whiteboard@1426183 13f79535-47bb-0310-9956-ffa450edef68